+++ lib/domain/factories/screen_factory.dart
@@ -13,6 +13,7 @@ import 'package:comics_db_app/ui/widgets/movie_details/components/trailer_widget
 import 'package:comics_db_app/ui/widgets/movie_details/movie_details_cubit.dart';
 import 'package:comics_db_app/ui/widgets/movie_details/movie_details_widget.dart';
 import 'package:comics_db_app/ui/widgets/movie_list/components/top_rated_movie_widget.dart';
+import 'package:comics_db_app/ui/widgets/movie_list/components/upcoming_movie_widget.dart';
 import 'package:comics_db_app/ui/widgets/movie_list/movie_list_cubit.dart';
 import 'package:comics_db_app/ui/widgets/movie_now_playing_list/movie_now_playing_list_widget.dart';
 import 'package:comics_db_app/ui/widgets/movie_now_playing_list/now_playing_movie_list_cubit.dart';
@@ -119,9 +120,11 @@ class ScreenFactory {
   Widget makeUpcomingMovieList() {
     return BlocProvider(
       create: (_) => UpcomingMovieListCubit(
-        upcomingMovieListBloc: UpcomingMovieListBloc(MovieListState.initial()),
+        upcomingMovieListBloc: UpcomingMovieListBloc(
+          MovieListState.initial(),
+        ),
       ),
-      child: const MovieNowPlayingListWidget(),
+      child: const UpcomingMovieWidget(),
     );
   }
 }