+++ lib/ui/widgets/people_details/components/people_top_poster_widget.dart
@@ -3,6 +3,7 @@ import 'package:comics_db_app/app_colors.dart';
 import 'package:comics_db_app/domain/api_client/image_downloader.dart';
 import 'package:comics_db_app/resources/resources.dart';
 import 'package:comics_db_app/ui/components/custom_poster_top_left_text.dart';
+import 'package:comics_db_app/ui/components/custom_poster_top_left_text_rating.dart';
 import 'package:comics_db_app/ui/components/loading_indicator.dart';
 import 'package:comics_db_app/ui/widgets/people_details/people_details_cubit.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
@@ -30,19 +31,20 @@ class PeopleTopPosterWidget extends StatelessWidget {
     return Stack(
       children: [
         // TODO delete this pic
+        profilePath != null ?
         Positioned(
           child: Opacity(
-            opacity: 0.01,
+            opacity: 0.001,
             child: AspectRatio(
               aspectRatio: 390 / 220,
               child: CachedNetworkImage(
-                imageUrl: ImageDownloader.imageUrl(profilePath!),
+                imageUrl: ImageDownloader.imageUrl(profilePath),
                 placeholder: (context, url) => const LoadingIndicatorWidget(),
-                errorWidget: (context, url, dynamic error) => Image.asset(AppImages.noImageBig),
+                errorWidget: (context, url, dynamic error) => Image.asset(AppImages.noImageAvailable),
               ),
             ),
           ),
-        ),
+        ) : const SizedBox.shrink(),
         Positioned(
           top: 45,
           left: 20,
@@ -96,36 +98,15 @@ class PeopleTopPosterWidget extends StatelessWidget {
                 const SizedBox(height: 7.0),
                 Row(
                   children: [
-                    CustomPosterTopLeftAlignText(text: gender == 1 ? female : male),
+                    CustomPosterTopLeftAlignText(text: gender == 1 ? female : male)
                   ],
                 ),
-                const SizedBox(height: 2.0),
                 Row(
                   children: [
-                    Align(
-                      alignment: Alignment.topLeft,
-                      child: Text(
-                        birthday ?? 'no birthday',
-                        style: const TextStyle(
-                          fontSize: 13,
-                          color: AppColors.genresText,
-                        ),
-                      ),
-                    ),
+                    CustomPosterTopLeftAlignText(text: birthday ?? 'no birthday'),
                   ],
                 ),
-                Align(
-                  alignment: Alignment.topLeft,
-                  child: Text(
-                    placeOfBirth ?? 'no place of birth',
-                    maxLines: 3,
-                    style: const TextStyle(
-                      fontSize: 13,
-                      color: AppColors.genresText,
-                    ),
-                  ),
-                ),
-                const SizedBox(height: 2),
+                CustomPosterTopLeftAlignText(text: placeOfBirth ?? 'no place of birth'),
                 Row(
                   children: [
                     const Icon(
@@ -133,15 +114,8 @@ class PeopleTopPosterWidget extends StatelessWidget {
                       color: AppColors.ratingStar,
                       size: 14,
                     ),
-                    const SizedBox(width: 4),
-                    Text(
-                      // popularity.toString(),
-                      popularity.toInt().toString(),
-                      style: const TextStyle(
-                        fontSize: 13,
-                        color: AppColors.ratingText,
-                      ),
-                    ),
+                    const SizedBox(width: 4.0),
+                    CustomPosterTopLeftAlignTextRating(text: popularity.toInt().toString()),
                   ],
                 ),
                 const SizedBox(height: 2),
@@ -149,8 +123,8 @@ class PeopleTopPosterWidget extends StatelessWidget {
             ),
           ),
         ),
+        profilePath != null ?
         Positioned(
-          // top: 55,
           left: 122,
           child: SizedBox(
             //     TODO: не закругляются края
@@ -159,10 +133,10 @@ class PeopleTopPosterWidget extends StatelessWidget {
             child: CachedNetworkImage(
               imageUrl: ImageDownloader.imageUrl(profilePath),
               placeholder: (context, url) => const LoadingIndicatorWidget(),
-              errorWidget: (context, url, dynamic error) => Image.asset(AppImages.noImageBig),
+              errorWidget: (context, url, dynamic error) => Image.asset(AppImages.noImageAvailable),
             ),
           ),
-        ),
+        ) : const SizedBox.shrink(),
       ],
     );
   }