+++ CHANGELOG.md
@@ -1,3 +1,15 @@
-0.0.1
+0.0.2
+--------------------
+- Added google fonts.
+- Fixed upcoming movies in navigation.
+- Changed "no image available" pic.
+- Changed `Image.Asset` to `CachedNetworkImage` at People screen.
+- Fixed if image is null.
+- Fixed people details `dateOfDeath` json.
+- Fixed `rendeflex` error at People cast widget.
+- Updated secure storage package.
+
+
+## 0.0.1
 ______________________
 - Initial release.